refactor(rimap-content): extract parse/pipeline.rs from parse/mod.rs
Moves parse_message + the six MIME-parse hard-limit constants
(MAX_MESSAGE_BYTES, MAX_BODY_BYTES, MAX_TOTAL_BODY_BYTES,
MAX_HEADER_BYTES, MAX_MIME_DEPTH, MAX_MIME_PARTS, MAX_HEADER_COUNT)
plus the tests module from parse/mod.rs (998 LOC) to a new
parse/pipeline.rs sibling (938 LOC). parse/mod.rs shrinks to 22 LOC
of pure wiring + re-exports.

mod.rs re-exports the constants and parse_message so existing
intra-dir `use crate::parse::MAX_*` and `use crate::parse::parse_message`
import sites (in attachments.rs, bodies.rs, meta.rs, plus the
external rimap-server import) continue to resolve. No public-API
change.

No behavior change. All 1416 workspace tests still pass.

Refs #300 (item 3, hub 2/4)
